requirements.txt
streamlit
faiss-cpu
numpy
python-dotenv
huggingface_hub
sentence-transformers
tiktoken